Refactored case adding view
