+++ taudinpurkauspeli/frontend/src/tests/case/NewCaseForm.test.js
@@ -3,7 +3,7 @@ import React from 'react';
 import '@testing-library/jest-dom/extend-expect';
 import { render, screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import NewCase from '../../components/case/NewCase';
+import NewCase from '../../components/case/NewCaseForm';
 
 jest.mock('react-i18next', () => ({
   useTranslation: () => ({ t: (key) => key }),
@@ -12,7 +12,7 @@ jest.mock('react-i18next', () => ({
 test('<AddCase /> updates parent state and calls onSubmit', async () => {
   const addCase = jest.fn();
 
-  render(<NewCase addCaseFunc={addCase} />);
+  render(<NewCase addCase={addCase} />);
 
   userEvent.type(screen.getByLabelText(/caseTitle/i), 'testTitle');
   userEvent.type(screen.getByLabelText(/caseAnamnesis/i), 'testAnamnesis');
@@ -29,7 +29,7 @@ test('<AddCase /> updates parent state and calls onSubmit', async () => {
 test('Case with invalid name cannot be created', async () => {
   const addCase = jest.fn();
 
-  render(<NewCase addCaseFunc={addCase} />);
+  render(<NewCase addCase={addCase} />);
 
   userEvent.type(screen.getByLabelText(/caseTitle/i), 't');
   userEvent.type(screen.getByLabelText(/caseAnamnesis/i), 'testAnamnesis');
